boxcxx/Ф23-backlog: error.h kernel-sync (X-macro single-source) + >2GiB int64 bridge + drift-guard ⇒ Ф23 100%
Closes the two Ф23 backlog items, taking Ф23 to true 100%, and guards the
single-source guarantee against future drift.

B-1 — boxlib↔kernel error.h sync via a single source of truth.
boxlib box/error.h was a drifted SUBSET (~100 of 202 kernel codes missing),
so a real kernel cause (e.g. ERR_ROUTE_NO_SUBSCRIBERS=941) reached box::error
nameless and printed "box:941 error" — the very errno-as-language remnant Ф23
set out to kill. The drift was structural: 3 hand-maintained userspace lists
(boxlib #defines, box::errc, box_error.cpp message+category) against the kernel.
Replaced with ONE BOX_ERROR_LIST(X) table in box/error.h carrying
{SUFFIX, name, value, "message", "category"} per row; the boxlib codes (now an
anonymous enum, as the kernel already is), box::errc, and both box_error.cpp
switch tables all expand from it. 3→1, no future re-drift. Verified 202/202
exact (name,value) parity with the kernel. The stale range-ladder category
function is deleted — each code carries its own category, so a boundary can
never drift again. Categories are maximally granular per owner (routing split
from ipc; TagFS modules split diskbook/cow/dedup/selfheal/boxhash/braid).

Active bug fixed: boxlib ERR_POCKET_FAILED was a wrong #define alias to
ERR_POCKET_PROCESSING_FAILED (906); the kernel's real ERR_POCKET_FAILED is 950.
Now a proper table row at 950. Zero userspace callers encoded the old 906
(grep-verified); ERR_POCKET_PROCESSING_FAILED(906) stays distinct. Locked by a
compile-time static_assert in Phase47.

B-2 — >2 GiB int-truncation in the fread/fwrite bridge.
fread/fwrite returned int; a 2..4 GiB byte count sign-truncated to negative,
which the box::result from_ret bridge mis-read as a -error_t cause; and
(uint32_t)size silently dropped size>=4GiB. Widened fread/fwrite to int64_t
(count 0..4GiB stays clear of the -error_t cause channel, every error_t<=1100)
and clamp the per-call request to UINT32_MAX (explicit short transfer — the
STORAGE wire field is 32-bit — instead of a silent wrap). current.c clamps to
INT32_MAX to keep its int-returning channel contract value-safe. New
box::_detail::from_ret64<T>(int64_t); box::tagfs read_at/write_at use it.

Drift-guard — tools/check_error_parity.sh, wired into `make all`.
The X-macro kills 3-way drift WITHIN userspace, but cannot reach the kernel
seam (boxlib is userspace-only). This guard diffs the kernel error_t enum
against BOX_ERROR_LIST and FAILS THE BUILD on any name/value mismatch, so a new
kernel code can never silently re-open the subset gap. Pure source diff,
instant; modelled on check_endbr64.sh. Self-tested (positive/negative/restore).

Test: Phase47 (cxxtest) — 941 names itself ("routing:941 no subscribers for
route"); a sweep proves every kernel range is categorized + named (no
box/unknown fallback); static_asserts lock POCKET_FAILED=950 and the
errc==ERR_* single-source guarantee; from_ret64 carries 3e9 as success and
recovers causes; storage-gated round-trip + read_object short-transfer→io
(degrades to a note, never a false PASS).

Gate: build clean, CET ENDBR64 1337/0, error-parity 202/202 (build-wired),
STRICT matrix BIOS/UEFI×{1c,16c} 4/4 ALL-PASS, Phase47 PASS 0 failures, 2
audits (c-pro SHIP + cpp-pro SHIP-WITH-FIXES; stale doc comments — "~70 codes",
"value 106", "macros" — corrected to match the now-true single-source reality).
Userspace-only; kernel error.h/error.c untouched (read as the source of truth).

+++ src/userspace/boxlib/include/box/file.h
@@ -27,8 +27,10 @@ typedef struct {
 int create(const char* filename, const char* tags);
 int query(const char* tags, uint32_t* file_ids, size_t max_files);
 int file_info(uint32_t file_id, file_info_t* info);
-int fread(uint32_t file_id, uint64_t offset, void* buffer, size_t size);
-int fwrite(uint32_t file_id, uint64_t offset, const void* buffer, size_t size);
+/* Return the byte count transferred (0..4 GiB, short-transfer: a request > 4 GiB
+ * is capped to one syscall — loop for more), or a negative -error_t on failure. */
+int64_t fread(uint32_t file_id, uint64_t offset, void* buffer, size_t size);
+int64_t fwrite(uint32_t file_id, uint64_t offset, const void* buffer, size_t size);
 /* `delete` is a C++ keyword — C++ callers use file_delete(), bound to the
  * same ELF symbol via asm label. C keeps the original name unchanged. */
 #ifdef __cplusplus
